moved update_jsones code into utils.py
threw this delinquent and questionable piece of artwork into another file.
